# [SPARK-46173] Skipping trimAll call during date parsing