--- a/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/DateTimeUtilsSuite.scala
+++ b/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/DateTimeUtilsSuite.scala
@@ -129,6 +129,9 @@ class DateTimeUtilsSuite extends SparkFunSuite with Matchers with SQLHelper {
 
   test("SPARK-32559: string to date trim Control Characters") {
     assert(toDate("MIDDLE\u0003") === None)
+    Seq("2015-03-18", "\u0003", "\u0003", " ", " ").permutations.map(p => p.mkString).foreach {
+      s => assert(toDate(s).get === days(2015, 3, 18))
+    }
   }
 
   test("string to date") {
@@ -155,12 +158,6 @@ class DateTimeUtilsSuite extends SparkFunSuite with Matchers with SQLHelper {
     assert(toDate("").isEmpty)
     assert(toDate("   ").isEmpty)
     assert(toDate("xxx2015-01-28").isEmpty)
-
-    // This is a problem:
-    // assert(toDate("2015-01-28 x").get === days(2015, 1, 28))
-    assert(toDate("2015-01-28 x").isEmpty)
-
-    assert(toDate("2015-01-28xx").isEmpty)
   }
 
   test("SPARK-35780: support full range of date string") {