--- a/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/DateTimeUtilsSuite.scala
+++ b/sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/DateTimeUtilsSuite.scala
@@ -128,10 +128,21 @@ class DateTimeUtilsSuite extends SparkFunSuite with Matchers with SQLHelper {
   }
 
   test("SPARK-32559: string to date trim Control Characters") {
-    assert(toDate("MIDDLE\u0003") === None)
-    Seq("2015-03-18", "\u0003", "\u0003", " ", " ").permutations.map(p => p.mkString).foreach {
-      s => assert(toDate(s).get === days(2015, 3, 18))
+    Seq("2015-03-18", "2015-03-18T123321", " 2015-03-18 123321", "+2015-03-18")
+      .foreach {
+        input => Seq(input, "\u0003", "\u0003", " ", " ")
+          .permutations.map(p => p.mkString).foreach {
+          s => assert(toDate(s).get === days(2015, 3, 18))
+        }
     }
+
+    Seq("INVALID_INPUT", " ", "1999-08-", "2015-03-18\u0003123321", "2015-03-18Q123321")
+      .foreach {
+        input =>
+          Seq(input, "\u0003", "\u0003", " ", " ").permutations.map(p => p.mkString).foreach {
+            s => assert(toDate(s).isEmpty)
+          }
+      }
   }
 
   test("string to date") {