@MaxGekk 2023-12-01
@dbatomic Could you remove this Link to [JIRA](https://issues.apache.org/jira/browse/SPARK-46173) from PR's description (PR's title has it already), and add the `[SQL]` tag to the title.