@MaxGekk 2023-12-02
> The change also includes a small unit benchmark for this particular case.

I wonder of other benchmarks. Do you observe perf regressions? I am asking just in case.