@dbatomic 2023-12-04
> > The change also includes a small unit benchmark for this particular case.
> 
> I wonder of other benchmarks. Do you observe perf regressions? I am asking just in case.

I just ran Parsing benchmarks locally. Here are the results:

With trimming changes:
func_name:                                Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
to timestamp str wholestage on                       85             87           2         11.7          85.2       1.1X
to_timestamp wholestage on                          421            423           3          2.4         420.7       1.0X
to_unix_timestamp wholestage on                     418            423           4          2.4         417.9       1.0X
to date str wholestage on                            98            101           1         10.2          98.3       0.9X
to_date wholestage on                               612            616           4          1.6         611.8       1.0X
to_date_with_trimming                               758            763           7          0.0    75792262.5       1.0X
to_date_with_trimming_only_suffix                     0              1           0          0.0       44237.5    1713.3X

--
Without my change
func_name:                                Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
to timestamp str wholestage on                       82             83           1         12.2          82.1       1.2X
to_timestamp wholestage on                          418            421           2          2.4         418.1       0.9X
to_unix_timestamp wholestage on                     414            416           2          2.4         413.9       0.9X
to date str wholestage on                           101            103           2          9.9         101.4       0.9X
to_date wholestage on                               620            624           4          1.6         620.3       1.0X

TLDR - no observable change. One some functions it is <1% worse and somewhere it is 1% better, but again, these are just tests on my local machine, without any real isolation so tiny diff is expected.