@dbatomic 2023-12-04
> > Looking for feedback if we want to keep the benchmark, given that this is a rather esoteric edge case.
> 
> I am ok to exclude the benchmark for the pretty specific case from the PR.

I decided to exclude the benchmark - thinking is that there is no point in having a benchmark for such an edge case.