@MaxGekk 2023-12-07
@dbatomic Does this suffer from the same issue?
https://github.com/apache/spark/blob/b4d90dd2a622beda542a7ce1ee15af9f312f9724/sql/api/src/main/scala/org/apache/spark/sql/catalyst/util/SparkDateTimeUtils.scala#L429

If so, could you optimize it too.